experiments/fitness-persona-generator/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
jinja2==3.1.4
pydantic==2.10.4
google-generativeai==0.8.3
python-multipart==0.0.20
